Move selection manager types into browser
